Feeds implementation with support for on-demand, continuous and timeout modes.
Feeds, in accordance to SPECS, are intermediary processes that are not
directly accessible to the HTTP clients, but instead serve as a proxy
connecting the AV source with the HTTP stream. Their purpose is to enable
multi-client use of a source that normally allows only a single-client
access.

As with streams, feeds can utilize any shell command that is capable
of outputting to stdout.

Feeds can be activated 'on-demand' when first HTTP stream client appears
where they are terminated once the last HTTP client disconnects. They can
also run in a 'continuous' mode, where they also start when the first
HTTP client appears, but don't stop upon final disconnect. Finally,
there is also a timeout mode in which the feed will be running for
a specified amount of seconds and only terminate if no new HTTP client
appears in this time.

Example configuration files (camera_to_mjpeg_*) have been prepared
to showcase each feed mode.