wip: feed-stream-client piping is now done in a single daemon thread
This allows to have a single point of call for select() on *all*
read and write pipes that currently exist within the process.

All pipes that report as ready can be handled in a single go, however
this also means this single go can transfer at most PIPE_BUF data
before it goes through maintenance routines and rewinds to the next
select() call. Queueing more than PIPE_BUF for reading may result in
memory overflow in the internal buffers. Queueing more for writing may
result in blocked pipes, though this last one is to be verified, yet.