Profile with yappi if VOYANDZ_PROFILE envvar is set
yappi profiler can be run by setting VOYANDZ_PROFILE environment
variable. This variable should be set to a file path where yappi's
pstats will be saved. Next to this file, with '.threads' extension,
thread information will also be dumped.

yappi is imported only if profiling mode is imported, therefore this
module is completely optional and should not be normally installed.