Feed mode is a piping.Mode enum now; create 'util' module
'util' module will contain boiler-plate code that can be used anywhere.
In this commit it defines a NameEnum enum that auto-translates its values
to lower-case strings and also provides methods to translate from Enum to
the string representation. This enum type is mainly supposed to handle
translation of limited-value settings from configuration to the program
internals.

monotonic() function has also been moved to the 'util' module.