Write only as much to the pipe as there's space left in it
Pipes, oh pipes, how difficult are you. The problem here is not in what
we can control but in what we can't control and also the perpetrators
cannot even be accused as their behavior is arguably not faulty.

When we do select() to see if FDs are write-ready, we do receive the FDs
that are ready correctly, however we still don't know how much data we
can write to them. If FD is a pipe, it has a limited size beyond which
no more data can be written to it until the receiving end reads it.
How do we know, however, if the receiving end flushed the whole FD,
just consumed the data partially or hasn't even touched it yet?
In `man 7 pipe` we can find the following quote:

> an application should be designed so that a reading process
> consumes data as soon as it is available

But assuming such a thing about all the applications in the world only
means that *you will get stuck* as you enivetably encounter a program
that doesn't follow this recommendation. We don't have the luxury of
getting stuck as this kills the whole server.

So, we need to prevent getting stuck. How do we do this? There are
several proposals:

1. Open pipes in O_NONBLOCK mode. Write-ends of such pipes will allow
   partial writes. select() will still await for pipes to become ready
   and os.write() will return the amount of bytes written, trying to
   write as much as possible. You can then buffer the remaining data
   and write it on the next iteration. This needs to be approached
   carefully as programs can mishandle such pipes. `cat` dies on
   a nonblocking read pipe. However, it's possible to set O_NONBLOCK flag
   only on the pipes used for internal data pushing and this seems to
   solve the problems with getting "stuck". *This is a viable solution.*

2. Write stuff one-byte-at-a-time. See https://lkml.org/lkml/2012/7/6/495
   Alternatively, write stuff 4096-bytes-at-a-time, however due to
   the size that needs to be pushed when pushing uncompressed data
   makes 4096 bytes only slightly less worse than one-byte-at-a-time.
   *This solution is not viable.*

3. By using ioctls it's possible to retrieve the total size of the pipe
   (`F_GETPIPE_SZ`) and the amount of data currently on the queue in
   the pipe (`FIONREAD`). We can then subtract the two values to see
   how much space there is left. Then, `os.write()` can write only as
   much data as there is space. *This is also a viable solution.*

This commit goes with solution no. 3.