Determine _PipeBuffer timeout as a delta between write and flush
It is unfair to determine that _PipeBuffer has timed out when nothing
has been written to it in the first place.

It is expected that the pipe buffers will buffer everything and then
flush it out to the other end. If we keep writing new data but the flush
gets stuck, the timeout counter properly kicks in. But if we're not
writing anything to the buffer, there's nothing to flush and thus
there's no way to establish if the flush is stuck. It's just as possible
that the writing will eventually resume and then continue normally.