Use setuptools-scm to dynamically determine the version from git tag
This is in accordance to SPECS.md where it is stated that the
development versions, made from commits between the tags, should have
their own version numbers indicating the commit.

Because .eggs directory can now be created, add a `distclean` target
to the Makefile that removes it.

Moreover, a lot more files now go into the sdist package. This is
absolutely correct, because the sdist package should contain the
entire source-code of the application. The MANIFEST.in still decides
what goes into the wheel.

The version known at runtime is no longer hardcoded in the version.py
file, but instead is extracted using importlib.metadata (or its backport
importlib_metadata in Python <3.8.1). The VERSION symbol is defaulted
in version.py to "V.O.Y", indicating an error if it's not replaced with
a proper version number at runtime.

The version attribute can be removed from setup.cfg as it's now
dynamically determined using setuptools-scm.